+++ nava/functions.py
@@ -5,17 +5,46 @@ import subprocess
 import os
 import shlex
 from functools import wraps
-import threading
-
+from .thread import NavaThread
 from .params import OVERVIEW
 from .params import SOUND_FILE_PLAY_ERROR, SOUND_FILE_EXIST_ERROR
-from .params import SOUND_FILE_PATH_TYPE_ERROR
+from .params import SOUND_FILE_PATH_TYPE_ERROR, SOUND_ID_EXIST_ERROR
 from .errors import NavaBaseError
+from . import params
 
-"""
-List of all aplay processes
-"""
-play_processes = []
+
+def stop(sound_id):
+    """
+    Stop sound.
+
+    :param sound_id: sound id
+    :type sound_id: int
+    :return: None
+    """
+    if sound_id not in params._play_threads_map:
+        raise NavaBaseError(SOUND_ID_EXIST_ERROR)
+    params._play_threads_map[sound_id].stop()
+
+
+def stop_all():
+    """
+    Stop all sounds.
+
+    :return: None
+    """
+    for thread in params._play_threads_map.values():
+        thread.stop()
+
+
+def sound_id_gen():
+    """
+    Sound id generator.
+
+    :return: sound id as int
+    """
+    params._play_threads_counter += 1
+    sound_id = params._play_threads_counter + 1000
+    return sound_id
 
 
 def nava_help():
@@ -53,137 +82,124 @@ def quote(func):
     return quoter
 
 
-def cleanup_processes():
-    """
-    Cleanup undead play processes after module exit.
-
-    :return: None
-    """
-    for proc in play_processes:
-        proc.kill()
-        proc.terminate()
-
-
-def __play_win(sound_path, is_async=True):
+def __play_win(sound_path, async_mode=False):
     """
     Play sound in Windows.
 
     :param sound_path: sound path
     :type sound_path: str
-    :param is_async: play async or not
-    :type is_async: bool
+    :param async_mode: async mode flag
+    :type async_mode: bool
+    :return: None or sound id
+    """
+    import winsound
+    play_flags = winsound.SND_FILENAME | (async_mode & winsound.SND_ASYNC)
+
+    if async_mode:
+        sound_thread = NavaThread(target=__play_win_flags,
+                                  args=(sound_path, play_flags), daemon=True)
+        sound_thread.start()
+        sound_id = sound_id_gen()
+        params._play_threads_map[sound_id] = sound_thread
+        return sound_id
+    else:
+        winsound.PlaySound(sound_path, play_flags)
+
+
+def __play_win_flags(sound_path, flags):
+    """
+    Play sound in Windows using different flags.
+
+    :param sound_path: sound path
+    :type sound_path: str
+    :param flags: different mode flags
+    :type flags: winsound flags
     :return: None
     """
     import winsound
-    # If is_async is ture, play async
-    play_flags = winsound.SND_FILENAME | (is_async & winsound.SND_ASYNC)
-    winsound.PlaySound(sound_path, play_flags)
+    winsound.PlaySound(sound_path, flags)
 
 
 @quote
-def __play_linux(sound_path, is_async=True):
+def __play_linux(sound_path, async_mode=False):
     """
     Play sound in Linux.
 
     :param sound_path: sound path to be played
     :type sound_path: str
-    :param is_async: play async or not
-    :type is_async: bool
-    :return: None or sound thread (depending on async flag)
+    :param async_mode: async mode flag
+    :type async_mode: bool
+    :return: None or sound id
     """
-    if is_async:
-        sound_thread = threading.Thread(target=__play_async_linux,
-                                        args=(sound_path,),
-                                        daemon=True)
+    if async_mode:
+        sound_thread = NavaThread(target=__play_proc_linux,
+                                  args=(sound_path,),
+                                  daemon=True)
         sound_thread.start()
-        return sound_thread
+        sound_id = sound_id_gen()
+        params._play_threads_map[sound_id] = sound_thread
+        return sound_id
     else:
-        __play_sync_linux(sound_path)
+        proc = __play_proc_linux(sound_path)
+        proc.wait()
 
 
-def __play_sync_linux(sound_path):
+def __play_proc_linux(sound_path):
     """
-    Play sound synchronously in Linux.
+    Create sound playing process in Linux.
 
     :param sound_path: sound path to be played
     :type sound_path: str
-    :return: None
+    :return: process
     """
-    _ = subprocess.check_call(["aplay",
-                            sound_path],
+    proc = subprocess.Popen(["aplay",
+                             sound_path],
                             shell=False,
                             stderr=subprocess.PIPE,
                             stdin=subprocess.PIPE,
                             stdout=subprocess.PIPE)
-    
-
-def __play_async_linux(sound_path):
-    """
-    Play sound asynchronously in Linux.
-
-    :param sound_path: sound path to be played
-    :type sound_path: str
-    :return: None
-    """
-    proc = subprocess.Popen(["aplay",
-                            sound_path],
-                            stderr=subprocess.PIPE,
-                            stdin=subprocess.PIPE,
-                            stdout=subprocess.PIPE)
-    play_processes.append(proc)
+    return proc
 
 
 @quote
-def __play_mac(sound_path, is_async=True):
+def __play_mac(sound_path, async_mode=False):
     """
     Play sound in macOS.
 
     :param sound_path: sound path
     :type sound_path: str
-    :param is_async: play sound in async mode
-    :type is_async: bool
-    :return: None or sound thread, depending on the flag
+    :param async_mode: async mode flag
+    :type async_mode: bool
+    :return: None or sound id
     """
-    if is_async:
-        sound_thread = threading.Thread(target=__play_async_mac,
-                                        args=(sound_path,),
-                                        daemon=True)
+    if async_mode:
+        sound_thread = NavaThread(target=__play_proc_mac,
+                                  args=(sound_path,),
+                                  daemon=True)
         sound_thread.start()
-        return sound_thread
+        sound_id = sound_id_gen()
+        params._play_threads_map[sound_id] = sound_thread
+        return sound_id
     else:
-        __play_sync_mac(sound_path)
+        proc = __play_proc_mac(sound_path)
+        proc.wait()
 
 
-def __play_sync_mac(sound_path):
+def __play_proc_mac(sound_path):
     """
-    Play sound synchronously in macOS.
+    Create sound playing process in macOS.
 
     :param sound_path: sound path to be played
     :type sound_path: str
-    :return: None
-    """
-    _ = subprocess.check_call(["afplay",
-                               sound_path],
-                              shell=False,
-                              stderr=subprocess.PIPE,
-                              stdin=subprocess.PIPE,
-                              stdout=subprocess.PIPE)
-
-
-def __play_async_mac(sound_path):
-    """
-    Play sound asynchronously in macOS.
-
-    :param sound_path: sound path to be played
-    :type sound_path: str
-    :return: None
+    :return: process
     """
     proc = subprocess.Popen(["afplay",
-                            sound_path],
+                             sound_path],
+                            shell=False,
                             stderr=subprocess.PIPE,
                             stdin=subprocess.PIPE,
                             stdout=subprocess.PIPE)
-    play_processes.append(proc)
+    return proc
 
 
 def path_check(func):
@@ -215,23 +231,23 @@ def path_check(func):
 
 
 @path_check
-def play(sound_path, is_async=True):
+def play(sound_path, async_mode=False):
     """
     Play sound.
 
     :param sound_path: sound path
     :type sound_path: str
-    :param is_async: play synchronously or asynchronously (async by default)
-    :type is_async: bool
-    :return: None or sound thread for futher handlings
+    :param async_mode: async mode flag
+    :type async_mode: bool
+    :return: None or sound id
     """
     try:
         sys_platform = sys.platform
         if sys_platform == "win32":
-            __play_win(sound_path, is_async)
+            return __play_win(sound_path, async_mode)
         elif sys_platform == "darwin":
-            return __play_mac(sound_path, is_async)
+            return __play_mac(sound_path, async_mode)
         else:
-            return __play_linux(sound_path, is_async)
-    except Exception: # pragma: no cover
+            return __play_linux(sound_path, async_mode)
+    except Exception:  # pragma: no cover
         raise NavaBaseError(SOUND_FILE_PLAY_ERROR)