+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `NavaThread` class
+- `stop` function
+- `stop_all` function
+### Changed
+- `async_mode` parameter added to `play` function
 ## [0.2] - 2023-07-10
 ### Added
 - Logo